% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSpower.R
\name{plot.PSpower}
\alias{plot.PSpower}
\title{Plots PSpower object}
\usage{
\method{plot}{PSpower}(x, power = seq(0.6, 0.99, length.out = 100), ...)
}
\arguments{
\item{x}{PSpower object}

\item{power}{a range of powers to plot the power curve}

\item{...}{ignored}
}
\value{
an object (class ggplot) containing a figure
}
\description{
Plots PSpower object
}
\examples{
obj <- PSpower(tau = 1/sqrt(20), sig.level = 0.05, power = 0.956,
               r = 0.5, phi = 0.99, rho_sq = 0.02)
plot(obj)
}
