% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlap.R
\name{plot_overlap}
\alias{plot_overlap}
\title{Plot density of propensity scores given treatment probability and overlap coefficient}
\usage{
plot_overlap(r, phi)
}
\arguments{
\item{r}{treatment probability}

\item{phi}{overlap coefficient}
}
\value{
a ggplot of the density of propensity scores in two treatment arms
}
\description{
Plot density of propensity scores given treatment probability and overlap coefficient
}
\examples{
plot_overlap(0.6, 0.9)
}
