% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSpower.R
\name{print.PSpower}
\alias{print.PSpower}
\title{Prints PSpower object}
\usage{
\method{print}{PSpower}(x, ...)
}
\arguments{
\item{x}{PSpower object}

\item{...}{ignored}
}
\description{
Prints PSpower object
}
