% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSpower.R
\name{plot.PSpower}
\alias{plot.PSpower}
\title{Plots PSpower object}
\usage{
\method{plot}{PSpower}(x, power = seq(0.6, 0.99, length.out = 100), ...)
}
\arguments{
\item{x}{PSpower object}

\item{power}{a range of powers to plot the power curve}

\item{...}{ignored}
}
\value{
an object (class ggplot) containing a figure
}
\description{
Plots PSpower object
}
\examples{
obj <- PSpower(1, 0.05, 0.956, 0.5, 0.99, -1.74, -2.74, 19.86, 20.12, 0.14, 0.14)
plot(obj)
}
